checker: fix signed integer literal overflow error, when most significant bit occupies signed bit (fix #23782) (#23919)

+++ vlib/rand/pcg32/pcg32_test.v
@@ -158,7 +158,7 @@ fn test_pcg32_u64_in_range() {
 
 fn test_pcg32_int31() {
 	max_u31 := int(0x7FFFFFFF)
-	sign_mask := int(0x80000000)
+	sign_mask := int(u32(0x80000000))
 	for seed in seeds {
 		mut rng := &rand.PRNG(&pcg32.PCG32RNG{})
 		rng.seed(seed)
@@ -174,7 +174,7 @@ fn test_pcg32_int31() {
 
 fn test_pcg32_int63() {
 	max_u63 := i64(0x7FFFFFFFFFFFFFFF)
-	sign_mask := i64(0x8000000000000000)
+	sign_mask := i64(u64(0x8000000000000000))
 	for seed in seeds {
 		mut rng := &rand.PRNG(&pcg32.PCG32RNG{})
 		rng.seed(seed)
